+++ Application/ResearchDataManagementPlatform/Menus/RDMPTopMenuStripUI.Designer.cs
@@ -61,6 +61,7 @@ namespace ResearchDataManagementPlatform.Menus
             this.listAllTypesToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.toolStripSeparator1 = new System.Windows.Forms.ToolStripSeparator();
             this.showPerformanceCounterToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+            this.lastCommandMonitorToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.openExeDirectoryToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.queryToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.queryCatalogue = new System.Windows.Forms.ToolStripMenuItem();
@@ -74,7 +75,11 @@ namespace ResearchDataManagementPlatform.Menus
             this.licenseToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.checkForUpdatesToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.rdmpTaskBar1 = new ResearchDataManagementPlatform.WindowManagement.TopBar.RDMPTaskBarUI();
-            this.lastCommandMonitorToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+            this.instancesToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+            this.switchToInstanceToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+            this.switchToDefaultSettings = new System.Windows.Forms.ToolStripMenuItem();
+            this.launchAnotherInstanceToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+            this.launchNewWithDefaultSettings = new System.Windows.Forms.ToolStripMenuItem();
             this.menuStrip1.SuspendLayout();
             this.SuspendLayout();
             // 
@@ -175,7 +180,8 @@ namespace ResearchDataManagementPlatform.Menus
             // 
             this.LocationsMenu.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
             this.configureExternalServersToolStripMenuItem,
-            this.setTicketingSystemToolStripMenuItem});
+            this.setTicketingSystemToolStripMenuItem,
+            this.instancesToolStripMenuItem});
             this.LocationsMenu.Name = "LocationsMenu";
             this.LocationsMenu.Size = new System.Drawing.Size(70, 20);
             this.LocationsMenu.Text = "Locations";
@@ -185,7 +191,7 @@ namespace ResearchDataManagementPlatform.Menus
             this.configureExternalServersToolStripMenuItem.Name = "configureExternalServersToolStripMenuItem";
             this.configureExternalServersToolStripMenuItem.ShortcutKeys = ((System.Windows.Forms.Keys)((System.Windows.Forms.Keys.Control | System.Windows.Forms.Keys.D)));
             this.configureExternalServersToolStripMenuItem.Size = new System.Drawing.Size(203, 22);
-            this.configureExternalServersToolStripMenuItem.Text = "Default Servers...";
+            this.configureExternalServersToolStripMenuItem.Text = "Change Default...";
             this.configureExternalServersToolStripMenuItem.Click += new System.EventHandler(this.configureExternalServersToolStripMenuItem_Click);
             // 
             // setTicketingSystemToolStripMenuItem
@@ -338,6 +344,13 @@ namespace ResearchDataManagementPlatform.Menus
             this.showPerformanceCounterToolStripMenuItem.Text = "Show Performance Counter...";
             this.showPerformanceCounterToolStripMenuItem.Click += new System.EventHandler(this.showPerformanceCounterToolStripMenuItem_Click);
             // 
+            // lastCommandMonitorToolStripMenuItem
+            // 
+            this.lastCommandMonitorToolStripMenuItem.Name = "lastCommandMonitorToolStripMenuItem";
+            this.lastCommandMonitorToolStripMenuItem.Size = new System.Drawing.Size(229, 22);
+            this.lastCommandMonitorToolStripMenuItem.Text = "Last Command Monitor";
+            this.lastCommandMonitorToolStripMenuItem.Click += new System.EventHandler(this.lastCommandMonitorToolStripMenuItem_Click);
+            // 
             // openExeDirectoryToolStripMenuItem
             // 
             this.openExeDirectoryToolStripMenuItem.Name = "openExeDirectoryToolStripMenuItem";
@@ -440,12 +453,46 @@ namespace ResearchDataManagementPlatform.Menus
             this.rdmpTaskBar1.Size = new System.Drawing.Size(1353, 29);
             this.rdmpTaskBar1.TabIndex = 57;
             // 
-            // lastCommandMonitorToolStripMenuItem
+            // instancesToolStripMenuItem
             // 
-            this.lastCommandMonitorToolStripMenuItem.Name = "lastCommandMonitorToolStripMenuItem";
-            this.lastCommandMonitorToolStripMenuItem.Size = new System.Drawing.Size(229, 22);
-            this.lastCommandMonitorToolStripMenuItem.Text = "Last Command Monitor";
-            this.lastCommandMonitorToolStripMenuItem.Click += new System.EventHandler(this.lastCommandMonitorToolStripMenuItem_Click);
+            this.instancesToolStripMenuItem.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
+            this.launchAnotherInstanceToolStripMenuItem,
+            this.switchToInstanceToolStripMenuItem});
+            this.instancesToolStripMenuItem.Name = "instancesToolStripMenuItem";
+            this.instancesToolStripMenuItem.Size = new System.Drawing.Size(203, 22);
+            this.instancesToolStripMenuItem.Text = "Instances";
+            // 
+            // switchToInstanceToolStripMenuItem
+            // 
+            this.switchToInstanceToolStripMenuItem.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
+            this.switchToDefaultSettings});
+            this.switchToInstanceToolStripMenuItem.Name = "switchToInstanceToolStripMenuItem";
+            this.switchToInstanceToolStripMenuItem.Size = new System.Drawing.Size(206, 22);
+            this.switchToInstanceToolStripMenuItem.Text = "Switch To Instance";
+            // 
+            // switchToDefaultSettings
+            // 
+            this.switchToDefaultSettings.Name = "switchToDefaultSettings";
+            this.switchToDefaultSettings.Size = new System.Drawing.Size(180, 22);
+            this.switchToDefaultSettings.Text = "Default";
+            this.switchToDefaultSettings.ToolTipText = "The RDMP instance recorded in your user settings";
+            this.switchToDefaultSettings.Click += switchToUsingUserSettings_Click;
+            // 
+            // launchAnotherInstanceToolStripMenuItem
+            // 
+            this.launchAnotherInstanceToolStripMenuItem.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
+            this.launchNewWithDefaultSettings});
+            this.launchAnotherInstanceToolStripMenuItem.Name = "launchAnotherInstanceToolStripMenuItem";
+            this.launchAnotherInstanceToolStripMenuItem.Size = new System.Drawing.Size(206, 22);
+            this.launchAnotherInstanceToolStripMenuItem.Text = "Launch Another Instance";
+            // 
+            // launchNewWithDefaultSettings
+            // 
+            this.launchNewWithDefaultSettings.Name = "launchNewWithDefaultSettings";
+            this.launchNewWithDefaultSettings.Size = new System.Drawing.Size(180, 22);
+            this.launchNewWithDefaultSettings.Text = "Default";
+            this.launchNewWithDefaultSettings.ToolTipText = "The RDMP instance recorded in your user settings";
+            this.launchNewWithDefaultSettings.Click += launchNewInstanceWithUserSettings_Click;
             // 
             // RDMPTopMenuStripUI
             // 
@@ -509,5 +556,10 @@ namespace ResearchDataManagementPlatform.Menus
         private System.Windows.Forms.ToolStripMenuItem newSessionToolStripMenuItem;
         private System.Windows.Forms.ToolStripMenuItem restartApplicationToolStripMenuItem;
         private System.Windows.Forms.ToolStripMenuItem lastCommandMonitorToolStripMenuItem;
+        private System.Windows.Forms.ToolStripMenuItem instancesToolStripMenuItem;
+        private System.Windows.Forms.ToolStripMenuItem launchAnotherInstanceToolStripMenuItem;
+        private System.Windows.Forms.ToolStripMenuItem launchNewWithDefaultSettings;
+        private System.Windows.Forms.ToolStripMenuItem switchToInstanceToolStripMenuItem;
+        private System.Windows.Forms.ToolStripMenuItem switchToDefaultSettings;
     }
 }